# Optimize block validation code and block propagation speed