--- a/src/limitedmap.h
+++ b/src/limitedmap.h
@@ -38,11 +38,11 @@ public:
     {
         std::pair<iterator, bool> ret = map.insert(x);
         if (ret.second) {
-            if (nMaxSize && map.size() == nMaxSize) {
+            rmap.insert(make_pair(x.second, ret.first));
+            if (nMaxSize && map.size() > nMaxSize) {
                 map.erase(rmap.begin()->second);
                 rmap.erase(rmap.begin());
             }
-            rmap.insert(make_pair(x.second, ret.first));
         }
         return;
     }