@gmaxwell 2015-02-27
WRT POW, I was worrying about the lose transaction case, not attacks in blocks.   The attack is the attacker has 1000 valid signatures, 1 invalid signature; the time is wasted on the valid signatures; the invalid is just there to avoid paying fees.  I don't see a way to cheaply invalidate the cache on the valid signatures.

The P2SH limit is per signature, I'm pretty sure you can get1000 checksigs in an IsStandard transaction, you'd just use 66 inputs with 15 checksigs each. With unusual script pubkeys (op_dup-ing the signatures) such a transaction would only be under15kb I'd guess.