@gavinandresen 2015-02-27
The 1,000-valid-plus-one-invalid attack will get the attacker DoS-banned for sending an invalid transaction.

They would have to Sybil and reconnect to get around that with either the old or new code; with the old code, they would have to pre-generate more valid signatures than the size of the signature cache AND Sybil to succeed in wasting CPU time. You're right, the new code makes the attack slightly easier.

But I think getting rid of the low-level cache is the right thing to do just because of the memory savings.