@gmaxwell 2015-02-27
Hm. So the caching isn't parametrized by the state of CCoinsview cache and validity can be conditional on it, but it's not clear to me that it matters.

What happens with this sequence of events.

Mempool contains transaction B that spends input X.
Block contains transactions A, B.   Both A and B spend X.   A wasn't in the mempool and will be a cache miss.  B is in the mempool, cached with inputs.HaveInputs(tx) passing.  But it should not pass in the context of a block containing tx A.   This would result in accepting blocks containing double spend pairs where half is in the mempool; but I believe it's currently prevented by a redundant test in ConnectBlock.

It would be good to have a test there, I could see this invariant being broken very easily.