@gmaxwell 2015-02-27
It appears that most of the performance comes from avoiding the /seemingly/ redundant interrogation of the CCoinsViewCache in CheckInputs:

'''
diff --git a/src/main.cpp b/src/main.cpp
index b6a61f7..6590b98 100644
--- a/src/main.cpp
+++ b/src/main.cpp
@@ -1436,11 +1436,6 @@ bool CheckInputs(const CTransaction& tx, CValidationState &state, const CCoinsVi
         if (pvChecks)
             pvChecks->reserve(tx.vin.size());
 
-        // This doesn't trigger the DoS code on purpose; if it did, it would make it easier
-        // for an attacker to attempt to split the network.
-        if (!inputs.HaveInputs(tx))
-            return state.Invalid(error("CheckInputs(): %s inputs unavailable", tx.GetHash().ToString()));
-
         // While checking, GetBestBlock() refers to the parent block.
         // This is also true for mempool checks.
         CBlockIndex *pindexPrev = mapBlockIndex.find(inputs.GetBestBlock())->second;
'''

Results in going from "Verify 1730 txins: 187.01ms (0.108ms/txin)" to "Verify 1730 txins: 13.29ms (0.008ms/txin)"

I've not yet convinced myself that its actually redundant, but if it is, it's good to take it out since it wastes time on mempool acceptance too.