@zyo2012 2015-02-27
Good job Gavin, this will help people complaining about the scalability of Bitcoin a bit. Every optimization bring down either the accessibility of running of full node or the scalability of whole system.