@gavinandresen 2015-02-27
@gmaxwell: I agree, a unit and/or regression test for one-double-spend-in-mempool, two-in-a-block is a very good idea.

RE: removing possibly redundant HaveInputs() :  should be a separate pull request. Thanks for reviewing and looking deeper!