@gmaxwell 2015-02-27
hah. I read it as bits to begin with (and was very confused by your most recent comment.