@SergioDemianLerner 2015-02-27
I responded to Gavin, but now that I see this thread, I copy parts here:

The sigcache was added for for security reasons and, as a by-product, it increases performance.

With a TxId cache you can achieve the performance part, but not the security protection.

The problem is described here https://bitcointalk.org/index.php?topic=136422.0 and only affects transactions during free forwarding (not in blocks).

To summarize the attack, the idea is that the attacker can send transactions that cost very little to create (can even reuse signatures with SIGHASH_ANYONECANPAY) and fail to verify after a lot of CPU has been wasted, but still the victim peer does not increase the DoS score of the attacker.  For example, re-using 999 good inputs, but failing in the 1000th one. 

The attacker can prevent DoS score to be increased because, there are two cases where it's not:
- If an input refers to a immature coinbase (even if it's not owned by the attacker)
- If an input has a not standard signature encoding.

So removing the sigcache has consequences.

Regarding performance, take into account that each signature is verified twice in AcceptToMemoryPool(), one with standard flags and another with mandatory flags (for extra security), so if you take out sigcache, then you double signature verification time.

Several solutions are possible:

1. Add a second level of cache for TxId's and get an additional speedup. This is probably the simplest and best.

2. Change VerifyInputs() so that it always increases the DoS score on a bad input, but the code explicitly states in one part:
   " Check whether the failure was caused by a non-mandatory script verification check ...; if so, don't trigger DoS protection to avoid splitting the network between upgraded and non-upgraded nodes. "

so you will be breaking this protection.

3. Add a DoS check that accumulates the amount of CPU wasted by a peer  (e.g. processing good inputs before a bad input) and increase the DoS score if the waste is over a threshold. Node resource encapsulation likes this is something Mike Hearn and I had pushed in the past, and I think M.H. is crowdfounding money for that purpose in his lighthouse project.

Obviously the best is just adding a second level of cache of TxIds, without changing anything security-related. I think its important though that the cache also caches the block number where the txId can be included, if the transaction uses locktime.