@sdaftuar 2015-02-27
Gavin, were you able to test the effect of this change on overall runtime?  Sergio's observation that signatures are checked twice on entry to the mempool seems like it could plausibly make things slower to take out the signature cache, and I've noticed in my initial testing that when I play back historical data through a node running this pull's code, I am seeing a substantial (nearly 2x) slowdown, although ConnectBlock is about 20% faster.